templated track counting tag computers
