+++ RecoBTag/ImpactParameter/interface/NegativeTrackCountingComputer.h
@@ -1,20 +1,11 @@
-#include "RecoBTag/ImpactParameter/interface/TrackCountingComputer.h"
+#ifndef ImpactParameter_NegativeTrackCountingComputer_h
+#define ImpactParameter_NegativeTrackCountingComputer_h
+
+#include "DataFormats/TrackReco/interface/TrackFwd.h"
+#include "DataFormats/BTauReco/interface/JTATagInfo.h"
+#include "RecoBTag/ImpactParameter/interface/TemplatedNegativeTrackCountingComputer.h"
 
 
-class NegativeTrackCountingComputer : public TrackCountingComputer
-{
- public:
-  NegativeTrackCountingComputer(const edm::ParameterSet & parameters ) : TrackCountingComputer(parameters)
-  {
-  }
- 
-  float discriminator(const TagInfoHelper & ti) const 
-   {
-     const reco::TrackIPTagInfo & tkip = ti.get<reco::TrackIPTagInfo>();
-     std::multiset<float> significances = orderedSignificances(tkip);
-     std::multiset<float>::iterator nth=significances.begin();
-     for(int i=0;i<m_nthTrack-1 && nth!=significances.end();i++) nth++;  
-     if(nth!=significances.end()) return -(*nth); else return -100.;
-   }
+typedef TemplatedNegativeTrackCountingComputer<reco::TrackRefVector,reco::JTATagInfo> NegativeTrackCountingComputer;
 
-};
+#endif // ImpactParameter_NegativeTrackCountingComputer_h
